ref: use dedicated highlight groups
Styling of the backdrop and comments is now configured using the
dedicated highlight groups `CommentHighlightsBackdrop` and
`CommentHighlightsComment`. These can be configured by the user or a
color theme.

Configuration via `opts.highlights` is now deprecated and will be
removed in a future release. Until then, it will still be supported in a
backwards-compatible manner.